requirements.txt
torch
ncps
scikit-learn
matplotlib
websocket-client
requests
streamlit
pandas
